집가기전 커밋 --소연님 userreservelist  페이징좀

+++ src/main/java/kr/ac/sungkyul/beautyline/dao/ReserveDao.java
@@ -21,7 +21,7 @@ public class ReserveDao {
 	}
 	
 	//회원 - 회원번호로 list 조회
-	public List<ReserveVo> resList( int userNo ){
+	public List<ReserveVo> resList( Long userNo ){
 		return sqlSession.selectList( "res.resListUserNo", userNo );
 	}
